user code support+ editor+
The user can now calculate and apply modifications manually without
having to fight the coroutine. Just set 'update' to false and call
'ModifyAll' or 'ModifyChunk' and then 'ApplyModifications' at your
leisure.

All of the code in ModifyAll, ModifyChunk and ApplyModifications used to
be in the AutoApplyModifiers coroutine (which used to be called
ApplyModifiers) but have now been moved into new methods to make useage
more modular.